services/ai-processing/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# AI相关
openai
langchain
langchain-openai
langchain-community

# 向量数据库
qdrant-client==1.7.0

# 数据库
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# Redis
redis==5.0.1

# HTTP客户端
httpx==0.26.0

# 工具
python-dotenv==1.0.0
python-multipart==0.0.6
faker==22.6.0
